% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_spec}
\alias{api_spec}
\title{API specification}
\usage{
api_spec()
}
\value{
A dataframe containing the endpoint and method names of supported
API methods. Additionally, information on whether paging is supported,
and which filter and sort values are allowed.
}
\description{
Retrieve a list of supported API methods and their supported dot arguments.
Each row in this dataframe has a corresponding function in the
\code{osmtags} package.
}
\examples{
api_spec()
}
