% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_chronology}
\alias{key_chronology}
\alias{key_combinations}
\alias{key_distribution}
\alias{key_overview}
\alias{key_prevalent_values}
\alias{key_projects}
\alias{key_similar}
\alias{key_stats}
\alias{key_values}
\alias{key_wiki_pages}
\title{Keys}
\source{
\url{https://taginfo.openstreetmap.org/taginfo/apidoc}
}
\usage{
key_chronology(key)

key_combinations(key, query = NULL, ...)

key_distribution(
  key,
  type = c("nodes", "ways"),
  path = tempfile(),
  plot = TRUE
)

key_overview(key)

key_prevalent_values(key, min_fraction = 0.01, ...)

key_projects(key, query = NULL, ...)

key_similar(key, query = NULL, ...)

key_stats(key)

key_values(key, lang = "en", query = NULL, ...)

key_wiki_pages(key)
}
\arguments{
\item{key}{Name of the key, e.g. \code{"amenity"} from \code{"amenity=hospital"}.}

\item{query}{A search term to filter results by. For \code{key_combinations}
and \code{key_similar}, this matches the \code{other_key} column, for
\code{key_projects} the project name or tag value, for \code{key_values}
the value. Defaults to \code{NULL}, i.e. no filtering.}

\item{...}{Additional generic parameters used for pagination, filtering,
and sorting. Can include the following:

\itemize{
 \item{\code{page}: If pagination is enabled, specifies the page number.}
 \item{\code{rp}: Enables pagination and specifies the number of results
 returned per page.}
 \item{\code{filter}: A character vector specifying the OSM geometry type
 to filter the results by. For example, if \code{"nodes"}, only tags on
 nodes are returned. Can be several of \code{"all"} (no filtering),
 \code{"nodes"}, \code{"relations"}, or \code{"ways"}.}
 \item{\code{sortname}: Numeric column of the output by which to arrange the
 results.}
 \item{\code{sortorder}: Whether to sort ascending (\code{"asc"}) or
 descending (\code{"desc"}).}
}
Note that not all additional arguments are defined for all API methods.
For details, refer to \code{\link{api_spec}}.}

\item{type}{OSM geometry type of the key whose distribution should be
plotted. Must be either \code{"nodes"} or \code{"ways"}. Defaults to
\code{"nodes"}.}

\item{path}{Path to a file where the plot should be written to.}

\item{plot}{Whether to plot the distribution using the R graphics device
using \code{\link[showimage]{show_image}}. If \code{FALSE}, simply returns
the path to the file. Defaults to \code{TRUE}.}

\item{min_fraction}{Numeric value specifying the fraction of all objects
that defines whether a value is prevalent. Defaults to 1 percent.}

\item{lang}{Language of the description text. Defaults to english.}
}
\value{
\code{tag_chronology}, \code{tag_combinations}, \code{tag_projects},
\code{tag_stats}, and \code{tag_wiki_pages} return a dataframe.
\code{tag_overview} returns a named list. \code{tag_distribution} returns
the path to an image and (optionally) plots the image using
\code{\link[showimage]{show_image}}.
}
\description{
Retrieve various data on OSM tags, i.e. key-value pairs. This includes:

\itemize{
 \item{\code{key_chronology}: Returns a history of nodes, ways, and
 relations for a given tag.}
 \item{\code{key_combinations}: Returns combinations of tags that often
 co-occur.}
 \item{\code{key_distribution}: Returns an image (or plot) of the spatial
 distribution of tags.}
 \item{\code{key_overview}: Returns a list containing a detailed summary
 of tag metadata.}
 \item{\code{key_prevalent_values}: Returns a list of the most prevalent
 values associated with a key, defined by a cut-off fraction.}
 \item{\code{key_projects}: Returns all projects using a tag.}
 \item{\code{key_similar}: Returns similar keys compared the specified key.}
 \item{\code{key_stats}: Returns the total number of nodes, ways, and
 relations associated with a tag.}
 \item{\code{key_values}: Returns all values used together with a key.}
 \item{\code{key_wiki_pages}: Returns information on the wiki pages
 associated with a tag.}
}
}
\examples{
\donttest{# Retrieve the history of a key
history <- key_chronology("amenity")
history$nodes <- cumsum(history$nodes)
with(history, plot(date, nodes, type = "l"))

# Only retrieve amenities that are explicitly buildings
key_combinations("amenity", query = "building")

# Only retrieve combinations for amenities that are mapped as relations
key_combinations("amenity", filter = "relations")

# Plot distribution of keys
key_distribution("amenity", type = "nodes")
key_distribution("highway", type = "ways")

summ <- key_overview("amenity")
str(summ)

# Retrieve all or just the prevalent values associated with a key
key_prevalent_values("hazmat")
key_values("hazmat")

# Retrieve JOSM projects on amenities
key_projects("amenity", query = "josm")

# Retrieve number of keys for each geometry type
key_stats("amenity")

# Retrieve wiki pages on a key
key_wiki_pages("amenity")}
}
