% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmtags-package.R
\docType{package}
\name{osmtags-package}
\alias{osmtags}
\alias{osmtags-package}
\title{osmtags: Tag statistics from OpenStreetMap}
\description{
Retrieves data on 'OpenStreetMap' tags using various instances of the 'taginfo' API.
}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} [copyright holder]

}
\keyword{internal}
