% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_key_value}
\alias{search_key_value}
\alias{search_keyword}
\alias{search_role}
\alias{search_value}
\title{Free search}
\source{
\url{https://taginfo.openstreetmap.org/taginfo/apidoc}
}
\usage{
search_key_value(query, ...)

search_keyword(query, ...)

search_role(query, ...)

search_value(query, ...)
}
\arguments{
\item{query}{Substring to search for. Used to search key/value combinations
in case of \code{search_key_value}, keywords in wiki pages in case of
\code{search_keyword}, relation roles in case of \code{search_role}, and
values in case of \code{search_value}}
}
\value{
For \code{search_key_value} and \code{search_keyword} returns
a dataframe containing the matched key-value combinations. For
\code{search_role} returns a dataframe containing the matched relation type
and role.
}
\description{
Search for tags, keys or relation roles.
}
\examples{
\donttest{search_key_value("highway=residential")
search_key_value("amenity=")
search_value("residential")

search_keyword("fire")
search_keyword("railway")

search_role("highway")}
}
