% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{taginfo_instances}
\alias{taginfo_instances}
\alias{set_taginfo_url}
\title{Taginfo instances}
\usage{
taginfo_instances(query = NULL, type = c("all", "wiki", "geofabrik"))

set_taginfo_url(name = NULL, url = NULL)
}
\arguments{
\item{query}{A search text that can be used to filter the description text.}

\item{type}{The source of the taginfo instance. \code{"wiki"} retrieves
all instances listed on the OSM wiki, and \code{"geofabrik"} retrives only
instances from Geofabrik. Returns all instances by default.}

\item{name}{Name of the instance to set. Corresponds to the \code{name}
column from \code{taginfo_instances()}. If both this and \code{url} are
\code{NULL}, falls back to the default API.}

\item{url}{A URL of a taginfo instance, e.g.
\code{"https://taginfo.openstreetmap.org/"}. If both this and \code{name} are
\code{NULL}, falls back to the default API.}
}
\value{
\code{taginfo_instances} returns a dataframe containing the source
of the instance info (\code{type}), the internal name, a human-readable
description, and the URL of the instance.
\code{set_taginfo_url} returns the new taginfo URL, invisibly.
}
\description{
By default, \code{osmtags} uses the official taginfo API. However, many
custom instances exist which can be retrieved and set using these functions.

\code{taginfo_instances} retrieves a number of known taginfo hosts from
the OSM wiki (\url{https://wiki.openstreetmap.org/wiki/Taginfo/Sites}) or
geofabrik (\url{https://taginfo.geofabrik.de/}). \code{set_taginfo_url}
changes the default API either by using the name listed in
\code{taginfo_instances()} or by directly providing the URL.
}
\examples{
# Get all instances for Africa
taginfo_instances("africa")

# Set South Africa as new taginfo API
set_taginfo_url("africa:south-africa")

# Reset taginfo API
set_taginfo_url()
}
