% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{unicode_chars}
\alias{unicode_chars}
\title{Parse unicode}
\source{
\url{https://taginfo.openstreetmap.org/taginfo/apidoc}
}
\usage{
unicode_chars(string)
}
\arguments{
\item{string}{A string.}
}
\value{
A dataframe containing details about each character including
the codepoint, script, category and character name.
}
\description{
Get information about a string of unicode characters.
}
\examples{
unicode_chars("a string")
}
