% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects_all}
\alias{projects_all}
\title{Projects}
\source{
\url{https://taginfo.openstreetmap.org/taginfo/apidoc}
}
\usage{
projects_all(query = NULL, status = "OK", ...)
}
\description{
Query projects.
}
